started implementation of SpeciesManager
